Prohibit assigning topology-based delivery services to caches (#4702)
* Prohibit assigning edges to delivery services if they have topologies

* Remove extra 'l', use log.Close

+++ traffic_ops/testing/api/v2/cachegroupsdeliveryservices_test.go
@@ -58,10 +58,10 @@ func CreateTestCachegroupsDeliveryServices(t *testing.T) {
 	}
 	cgID := *clientCG.ID
 
-	dsIDs := []int64{}
+	dsIDs := []int{}
 	for _, ds := range dses {
 		if *ds.CDNName == "cdn1" {
-			dsIDs = append(dsIDs, int64(*ds.ID))
+			dsIDs = append(dsIDs, *ds.ID)
 		}
 	}
 	if len(dsIDs) < 1 {
